// src/app/shared/models/error-constant.ts
export enum ErrorConstants {
  CONST_ERROR_INVALID_DEVICE_ID = 'error_const_invalid_device_id',
  CONST_ERROR_BUSINESS_CENTER_ID_INVALID = 'const_error_invalid_business_center',
  CONST_ERROR_PARTNER_ID_IS_NOT_ASSOCIATED_WITH_BUSINESS_CENTER = 'const_error_partner_is_not_associated_with_business_center',
  CONST_ERROR_BUSINESS_CENTER_HAS_NO_LOCATION = 'const_error_business_center_has_no_location',
  CONST_ERROR_DEVICES_NOT_FOUND = 'error_devices_not_found',
  CONST_ERROR_DEVICE_ID_MUST_BE_UNIQUE = 'const_error_device_id_must_be_unique',
  CONST_ERROR_POS_CLEINT_ID_MUST_BE_UUID = 'const_error_pos_client_id_must_be_uuid',
  CONST_ERROR_INVALID_POS_CLIENT_ID = 'const_error_invalid_pos_client_id',
  CONST_ERROR_NO_POS_CLIENTS_FOUND = 'const_error_no_pos_clients_found',
  CONST_ERROR_MENU_ITEMS_ID_MUST_BE_UUID = 'const_error_menu_items_must_be_uuid',
  CONST_ERROR_MENUITEMS_MUST_BE_ARRAY = 'const_error_menu_items_must_be_array',
  CONST_ERROR_MENU_ITEM_NOT_FOUND = 'error_menu_item_not_found',
  CONST_ERROR_MENUITEM_ALREADY_EXIST = 'error_menu_items_already_exist_for_a_device',
  CONST_ERROR_BUSINESS_CENTRE_HAS_NO_LOCATION = 'error_business_center_has_no_location',
  CONST_ERROR_SCHEDULED_DATE_IS_REQUIRED = 'error_schedules_date_is_required',
  CONST_ERROR_NO_ACTIVE_BC_OR_SCHEDULE_MENU = 'error_no_active_bc_or_schedule_menu',
}

export enum ScheduleErrorConstants {
  CONST_ERROR_SCHEDULE_MENU_DETAILS_NOT_FOUND = 'error_schedule_menu_details_not_found',
  CONST_ERROR_MENU_ITEM_DETAILS_NOT_FOUND = 'error_menu_item_details_not_found',
  CONST_ERROR_MENU_DETAILS_EXISTS = 'error_menu_details_already_exists',
  CONST_ERROR_INVALID_UPDATE_SCHEDULE_DATE = 'error_invalid_schedule_menu_date',
  CONST_ERROR_SCHEDULE_MENU_ALREADY_ACTIVE = 'error_schedule_menu_alredy_active',
  CONST_ERROR_SCHEDULE_MENU_COMPLETED = 'error_schedule_menu_already_completed',
  CONST_ERROR_SCHEDULE_MENU_ALREADY_DRAFT = 'error_schedule_menu_alredy_in_draft',
  CONST_ERROR_SCHEDULE_MENU_ALREADY_INACTIVE = 'error_schedule_menu_alredy_in_inactive',

}